Use versioning transition checks (#27)
